move datasource logos
