+++ src/pages/administration/datasources/index.vue
@@ -98,9 +98,9 @@ export default {
 		},
 		mapTypeToDatasourceImage(item) {
 			// todo later - check naming
-			const webuntis = require("@assets/img/logo/logo-webuntis.svg");
-			const ldap = require("@assets/img/logo/logo-ldap.svg");
-			const rss = require("@assets/img/logo/logo-rss.svg");
+			const webuntis = require("@assets/img/datasources/logo-webuntis.svg");
+			const ldap = require("@assets/img/datasources/logo-ldap.svg");
+			const rss = require("@assets/img/datasources/logo-rss.svg");
 			const mapping = { webuntis, ldap, rss };
 			return mapping[item.config.type];
 		},
